- aligns request options with configuration revamp
