+++ Microsoft.Kiota.Abstractions.Tests/RequestInformationTests.cs
@@ -39,7 +39,7 @@ namespace Microsoft.Kiota.Abstractions.Tests
             var testRequestOption = new Mock<IRequestOption>().Object;
             Assert.Empty(testRequest.RequestOptions);
             // Act
-            testRequest.AddRequestOptions(testRequestOption);
+            testRequest.AddRequestOptions(new IRequestOption[] {testRequestOption});
             // Assert
             Assert.NotEmpty(testRequest.RequestOptions);
             Assert.Equal(testRequestOption, testRequest.RequestOptions.First());
@@ -48,5 +48,46 @@ namespace Microsoft.Kiota.Abstractions.Tests
             testRequest.RemoveRequestOptions(testRequestOption);
             Assert.Empty(testRequest.RequestOptions);
         }
+        [Fact]
+        public void SetsSelectQueryParameters()
+        {
+            // Arrange as the request builders would
+            var requestInfo = new RequestInformation
+            {
+                HttpMethod = Method.GET,
+                UrlTemplate = "http://localhost/me{?%24select}"
+            };
+            Action<GetQueryParameters> q = x => x.Select = new[] { "id", "displayName" };
+            var qParams = new GetQueryParameters();
+            q.Invoke(qParams);
+
+            // Act 
+            requestInfo.AddQueryParameters(qParams);
+
+            // Assert
+            Assert.True(requestInfo.QueryParameters.ContainsKey("%24select"));
+            Assert.False(requestInfo.QueryParameters.ContainsKey("select"));
+            Assert.Equal("%24select",requestInfo.QueryParameters.First().Key);
+        }
+    }
+
+    /// <summary>The messages in a mailbox or folder. Read-only. Nullable.</summary>
+    internal class GetQueryParameters
+    {
+        /// <summary>Select properties to be returned</summary>\
+        [QueryParameter("%24select")]
+        public string[] Select { get; set; }
+        /// <summary>Include count of items</summary>
+        [QueryParameter("%24count")]
+        public bool? Count { get; set; }
+        /// <summary>Expand related entities</summary>
+        [QueryParameter("%24filter")]
+        public string Filter { get; set; }
+        /// <summary>Order items by property values</summary>
+        [QueryParameter("%24orderby")]
+        public string[] Orderby { get; set; }
+        /// <summary>Search items by search phrases</summary>
+        [QueryParameter("%24search")]
+        public string Search { get; set; }
     }
 }
